コンポーネントを細分化し、コード量を減らした
