+++ src/components/presentation/_EditArea/EditorDisplay.tsx
@@ -1,15 +1,5 @@
-import React, {
-  useState,
-  useEffect,
-  useMemo,
-  useCallback,
-  Fragment,
-  useRef,
-  useContext,
-  createContext,
-} from "react";
+import { useState, useEffect, useMemo, createContext, useRef } from "react";
 import { Flex } from "@chakra-ui/react";
-import { getCodeString } from "rehype-rewrite";
 import mermaid from "mermaid";
 
 import { DisplayModes, type DisplayMode } from "@/utils/constants";
@@ -18,8 +8,11 @@ import { useDebounce } from "@/utils/hooks/useDebounce";
 
 import MDEditor, { commands, type MDEditorProps } from "@uiw/react-md-editor";
 
+import { Code } from "@/components/parts/editor/editorDisplay/Code";
+import { AnchorTag } from "@/components/parts/editor/editorDisplay/AnchorTag";
+
 // themeを保持するコンテキスト
-const AppSettingContext = createContext<{
+export const AppSettingContext = createContext<{
   theme: AppTheme;
 }>({
   theme: AppThemes.SYSTEM,
@@ -30,72 +23,6 @@ export type EditorDisplayProps = {
   theme: AppTheme;
   displayMode: DisplayMode;
   saveMdText: (mdText: string) => void;
-};
-
-const randomid = () => parseInt(String(Math.random() * 1e15), 10).toString(36);
-const Code = ({
-  inline,
-  children = [],
-  className,
-  ...props
-}: {
-  inline?: boolean;
-  children?: string[];
-  className?: string;
-  node?: {
-    children: any[];
-  };
-}) => {
-  const { theme } = useContext(AppSettingContext);
-  const demoid = useRef(`dome${randomid()}`);
-  const [container, setContainer] = useState<HTMLElement | null>(null);
-  const isMermaid =
-    className && /^language-mermaid/.test(className.toLocaleLowerCase());
-  const code: string =
-    props.node && props.node.children
-      ? getCodeString(props.node.children)
-      : children[0] || "";
-
-  const reRender = async () => {
-    if (container && isMermaid) {
-      try {
-        const str = await mermaid.render(demoid.current, code);
-        container.innerHTML = str.svg;
-      } catch (error) {
-        container.innerHTML = String(error);
-      }
-    }
-  };
-
-  useEffect(() => {
-    reRender();
-  }, [container, isMermaid, code, demoid, theme]);
-
-  const refElement = useCallback((node: HTMLElement | null) => {
-    if (node !== null) {
-      setContainer(node);
-    }
-  }, []);
-
-  if (isMermaid) {
-    return (
-      <Fragment>
-        <code id={demoid.current} style={{ display: "none" }} />
-        <code ref={refElement} data-name="mermaid" />
-      </Fragment>
-    );
-  }
-  return <code className={className}>{children}</code>;
-};
-
-// 引用: https://qiita.com/wataru775/items/61db1371655897aea517 (感謝)
-const AnchorTag = ({ node, children, ...props }: any) => {
-  try {
-    new URL(props.href ?? "");
-    props.target = "_blank";
-    props.rel = "noopener noreferrer";
-  } catch (e) {}
-  return <a {...props}>{children}</a>;
 };
 
 export const EditorDisplay = ({
@@ -213,7 +140,7 @@ export const EditorDisplay = ({
         <MDEditor
           value={mdLocalText}
           onChange={handleChange}
-          height={"100%"}
+          height="100%"
           style={{
             width: "100%",
           }}
